backend/ai_server/requirements.txt
flask
flask-cors
transformers
torch
Pillow